Aligning with new features of JDK8

+++ extension/src/main/java/com/alibaba/fastjson2/support/AwtRederModule.java
@@ -72,7 +72,7 @@ public class AwtRederModule
         public Color apply(Map<Long, Object> values) {
             Integer rgb = (Integer) values.get(HASH_RGB);
             if (rgb != null) {
-                return new Color(rgb.intValue());
+                return new Color(rgb);
             }
 
             Integer r = (Integer) values.get(HASH_R);
